review: fold SM80 adapters into api.py, configs into config_sm80.py, drop the docs pages
Per maintainer feedback on the PR:
- SdpafwdSm80/SdpabwdSm80 + wrappers move from api_sm80.py into each opset's
  api.py, following the SM100 classes there (one api.py per opset, no per-arch
  files). Lazy exports and the engine lowerings repoint; no signature changes.
- The per-flavor kernel configs leave kernels/ for parent-level config_sm80.py
  (one per direction), matching config_sm100/config_sm120.
- The FE OSS docs pages and their overview/Attention/llms.txt entries are
  dropped for now; the engines remain the PR's product.
