+++ docs/fe-oss-apis/overview.md
@@ -34,8 +34,6 @@ This folder documents the Python FE APIs implemented under `python/cudnn`. For d
 - [SDPA Forward FE OSS API (SM100, D=256)](https://docs.nvidia.com/deeplearning/cudnn/frontend/latest/operations/Attention.html#sdpa-forward-fe-oss-sm100-d256)
 - [SDPA Backward FE OSS API (SM100, D=256)](https://docs.nvidia.com/deeplearning/cudnn/frontend/latest/operations/Attention.html#sdpa-backward-fe-oss-sm100-d256)
 - [SDPA Backward (SM120)](attention/sdpa_bwd_sm120.md)
-- [SDPA Forward FE OSS API (SM80, D<=256)](attention/sdpa_fwd_sm80.md)
-- [SDPA Backward FE OSS API (SM80, D<=256)](attention/sdpa_bwd_sm80.md)
 - [RMSNorm + SiLU](rmsnorm_silu.md)
 
 ## Installation and setup
