+++ python/cudnn/sdpa/fwd/__init__.py
@@ -21,8 +21,8 @@ _LAZY_EXPORTS = {
     "SdpaFwdDslSm120": (".api_dsl", "SdpaFwdDslSm120"),
     "sdpa_fwd_wrapper_dsl_sm100": (".api_dsl", "sdpa_fwd_wrapper_dsl_sm100"),
     "sdpa_fwd_wrapper_dsl_sm120": (".api_dsl", "sdpa_fwd_wrapper_dsl_sm120"),
-    "SdpafwdSm80": (".api_sm80", "SdpafwdSm80"),
-    "sdpa_fwd_wrapper_sm80": (".api_sm80", "sdpa_fwd_wrapper_sm80"),
+    "SdpafwdSm80": (".api", "SdpafwdSm80"),
+    "sdpa_fwd_wrapper_sm80": (".api", "sdpa_fwd_wrapper_sm80"),
 }
 
 __all__ = list(_LAZY_EXPORTS)
